requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
django-crispy-forms==1.14.0
django-js-asset==2.0.0
django-utils-six==2.0
gunicorn==20.1.0
Pillow==9.1.1
pytz==2022.1
sqlparse==0.4.2
whitenoise==6.1.0
